+++ python/fnllm/CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ```md
 ## [Unreleased]
 ### Added
+- Add `cache_bust` flag to LLMInput.
 ### Changed
 ### Deprecated
 ### Removed